Basic specs for `Rc` and `Arc`
