+++ creusot-contracts/src/std.rs
@@ -18,9 +18,11 @@ pub mod ops;
 pub mod option;
 pub mod panicking;
 pub mod ptr;
+pub mod rc;
 pub mod result;
 pub mod slice;
 pub mod string;
+pub mod sync;
 pub mod time;
 mod tuples;
 pub mod vec;
